Lot of rework to get CACHE running in special cases on Spartan 6 / ISE
The old implementaiton led to a a not working implementation of the
Cache, because the Cache RAMs where not wired correctly. This seems to
be an XST bug, because simulaiton and Synthesis with Vivado for Artix-7
is correct. The work around is use Xilinx BRAM Macros, but this is only
implemented for 2K*32Bit Cache size.
A side effect the code is refactored to separate the cache RAM from the
main chache module